remove t2hs.rb; move into lib/aozora2html.rb
