+++ .rubocop.yml
@@ -29,7 +29,7 @@ Style/FormatStringToken:
 
 Style/FrozenStringLiteralComment:
   Exclude:
-    - 'lib/t2hs.rb'
+    - 'lib/aozora2html.rb'
 
 Style/RedundantPercentQ:
   Enabled: false
